Adding explicit return type by casting to actual Type in each Shim lambda created by reflection (and thus giving the lambda a static return type instead of object, thus allowing ValueTypes to box properly; prior to these changes, int was being treated as object and losing its integer value when ToString was called).
Also added the ability to specify whether properties should be mocked or not for a particular Shimmy (a particular class).

Left calls in the ShimmyTests class for an example of workarounds for framework issues (with string.format or LINQ, for example). Will need to write this up in some sort of guide. The pattern is simple -- just call the actual framework method directly in the Shim -- but it would be good to have some explanation.

Also some cleanup, including enabling Shim With statements with lambdas to automatically resolve their recipient methods (those that must be shimmed).